+++ hms-sdk-unity/HuaweiMobileServices/Game/AntiAddictionCallWrapper.cs
@@ -1,0 +1,19 @@
+﻿using UnityEngine;
+
+namespace HuaweiMobileServices.Game
+{
+    public class AntiAddictionCallWrapper : AndroidJavaProxy
+    {
+        private readonly IAntiAddictionCallback mListener;
+
+        public AntiAddictionCallWrapper(IAntiAddictionCallback listener) : base("com.huawei.hms.jos.AntiAddictionCallback")
+        {
+            mListener = listener;
+        }
+
+        public void onExit()
+        {
+            mListener.OnExit();
+        }
+    }
+}
