+++ hms-sdk-unity/HuaweiMobileServices/Game/IAntiAddictionCallback.cs
@@ -1,0 +1,7 @@
+﻿namespace HuaweiMobileServices.Game
+{
+    public interface IAntiAddictionCallback
+    {
+        void OnExit();
+    }
+}
